+++ src/rotas.js
@@ -3,7 +3,7 @@
 const express = require('express');
 const rotas = express();
 const {listarContas, cadastrarConta, atualizarConta, deletarConta, depositar, sacar, transferir, saldo, extrato} = require('./controladores/controladores');
-const {validarSenhaBanco, validarEmailCpf, validarDados, validarContaBanco, validardeposito} = require('./intermediarios/intermediarios');
+const {validarSenhaBanco, validarEmailCpf, validarDados, validarContaBanco, validardeposito, validarSaque} = require('./intermediarios/intermediarios');
 
 
 
@@ -13,7 +13,7 @@ rotas.get('/contas/extrato', extrato)
 
 rotas.post('/contas', validarEmailCpf, validarDados, cadastrarConta);
 rotas.post('/transacoes/depositar',validardeposito, depositar);
-rotas.post( '/transacoes/sacar',  sacar);
+rotas.post( '/transacoes/sacar',validarSaque,  sacar);
 rotas.post('/transacoes/transferir',  transferir);
 
 rotas.put('/contas/:numeroConta',validarDados, validarEmailCpf, validarContaBanco,  atualizarConta);
